No need to drop database in CI
Spotted at https://github.com/Starkast/wikimum/runs/934679352?check_suite_focus=true#step:5:36

    dropdb: error: could not connect to database template1: could not connect to server: No such file or directory
      Is the server running locally and accepting
      connections on Unix domain socket "/var/run/postgresql/.s.PGSQL.5432"?

+++ test/test_database.rb
@@ -28,7 +28,7 @@ module TestDatabase
 
     DB.disconnect
 
-    system("dropdb #{database_name}")
+    system("dropdb #{database_name}") unless ENV["TEST_DATABASE_URL"]
   end
 end
